Refactor: Moved pipeline register modules to upper level.
Details: Changed pipeline register module's position in hierarchy from stage modules to datapath module.
This improves the scalability of the architecture if number of pipelines to be changed or other architectural
changes to be made. It makes it easiser to track which stage signal belongs. I changed signal names in
datapath module as well to realize it.